\name{check.equality}
\alias{check.equality}
\title{A small GUI for Computer Algebra within Lyx
Checks whether a latex equality holds true}
\usage{
  check.equality(txt)
}
\description{
  A small GUI for Computer Algebra within Lyx Checks
  whether a latex equality holds true
}

\examples{\dontrun{
  set.storing(TRUE)
  check.equality("x_{1}+x_{1}=2x_{1}")
  check.equality("x_{1}+x_{1}=x_{1}")
  
}}
