\name{ma.mao.simplify}
\alias{ma.mao.simplify}
\title{Takes maxima code and transforms it into a simplification command}
\usage{
  ma.mao.simplify(term, lyma = new.lyma())
}
\description{
  Takes maxima code and transforms it into a simplification
  command
}

\examples{\dontrun{
  ma.mao.simplify("x1+x1+x1")
}}
