\name{remove.redundant.maxima.braces}
\alias{remove.redundant.maxima.braces}
\title{Removes many redundant braces in a string str that contains a maxima mathematical expression}
\usage{
  remove.redundant.maxima.braces(str)
}
\description{
  Removes many redundant braces in a string str that
  contains a maxima mathematical expression
}

